+++ src/WebJobs.Extensions.DurableTask/Options/AzureStorageOptions.cs
@@ -18,6 +18,7 @@ namespace Microsoft.Azure.WebJobs.Extensions.DurableTask
         private const int MaxTaskHubNameSize = 45;
         private const int MinTaskHubNameSize = 3;
         private const string TaskHubPadding = "Hub";
+        private TimeSpan maxQueuePollingInterval;
 
         /// <summary>
         /// Gets or sets the name of the Azure Storage connection information used to manage the underlying Azure Storage resources.
@@ -160,9 +161,36 @@ namespace Microsoft.Azure.WebJobs.Extensions.DurableTask
 
         /// <summary>
         /// Gets or sets the maximum queue polling interval.
+        /// We update the default value to 1 second for the Flex Consumption SKU
+        /// because of a known cold start latency with Flex Consumption
+        /// and Durable Functions.
+        /// The default value is 30 seconds for all other SKUs.
         /// </summary>
         /// <value>Maximum interval for polling control and work-item queues.</value>
-        public TimeSpan MaxQueuePollingInterval { get; set; } = TimeSpan.FromSeconds(30);
+        public TimeSpan MaxQueuePollingInterval
+        {
+            get
+            {
+                if (this.maxQueuePollingInterval == TimeSpan.Zero)
+                {
+                    if (string.Equals(Environment.GetEnvironmentVariable("WEBSITE_SKU"), "FlexConsumption", StringComparison.OrdinalIgnoreCase))
+                    {
+                        this.maxQueuePollingInterval = TimeSpan.FromSeconds(1);
+                    }
+                    else
+                    {
+                        this.maxQueuePollingInterval = TimeSpan.FromSeconds(30);
+                    }
+                }
+
+                return this.maxQueuePollingInterval;
+            }
+
+            set
+            {
+                this.maxQueuePollingInterval = value;
+            }
+        }
 
         /// <summary>
         /// Determines whether or not to use the old partition management strategy, or the new